feat(box): make boxes and holders attachable to grids
Users can now attach boxes and holders to grids indicating its position
on the grid. If you move the grid all attached object move with it.

Translation of object is achived by attaching the XZ plane of the box to
the XZ plane of the grid. Then X and Y relative coordinates are updated
as multiples of 50[mm].

Usage:
- Select a grid on the tree
- Click on the box/holder command
- Update the PositionX/PositionY properties of the box/holder