fix(box/grid): set limits on integer and length properties
- Use `App::PropertyIntegerConstraint` instead of `App::PropertyInteger`
in order to get a smaller set of valid options for the parts.

Minimum, maximum, step and default values are set.

Previously you could enter negative numbers which were fixed at the time
of generating the shape avoiding problems but the properties were not
fixed. Now users are not able to enter bad data which leads to a
simplification of the sanitization code.

Proposed upper limit at 50. Seems sane value.

- Created method to check length properties are inside limits.